Extract massive inline tests to reduce Blob file tangles
Extracted inline tests using `#[cfg(test)] mod tests;` into dedicated `tests.rs` files within module directories (e.g., `algebra/summarize/tests.rs`) for multiple bloated 'God Files'. This reduces source sizes by up to 50%, improving code readability and cohesion without changing semantics.
